# Optimize Proto decimal conversion