<!--
Suggested PR template: Fill/delete/add sections as needed. Optionally delete any commented block.
-->
What
----
This PR optimizes the conversion between .NET decimals and Protobuf decimals by leveraging the latest API features and eliminating unnecessary heap allocations.

Before
| Method                      | Mean     | Error   | StdDev  | Gen0   | Allocated |
|---------------------------- |---------:|--------:|--------:|-------:|----------:|
| Benchmark_ToProtobufDecimal | 120.2 ns | 0.26 ns | 0.21 ns | 0.1376 |     288 B |
| Benchmark_ToSystemDecimal   | 183.7 ns | 0.31 ns | 0.28 ns | 0.1566 |     328 B |

After 
| Method                      | Mean      | Error    | StdDev   | Gen0   | Allocated |
|---------------------------- |----------:|---------:|---------:|-------:|----------:|
| Benchmark_ToProtobufDecimal |  44.91 ns | 0.692 ns | 0.647 ns | 0.0650 |     136 B |
| Benchmark_ToSystemDecimal   | 172.02 ns | 1.023 ns | 0.799 ns | 0.1414 |     296 B |


Checklist
------------------
* [ ] Contains customer facing changes? Including API/behavior changes
  * No
* [ ] Did you add sufficient unit test and/or integration test coverage for this PR?
  * Not required — behavior remains unchanged and is already covered.

cc: @rayokota